chore: rename create console script
