+++ app/shell/py/pie/pie/create/__init__.py
@@ -1,0 +1,6 @@
+"""Helpers for creating Press projects and posts."""
+
+__all__ = ["site", "post"]
+
+# Export submodules for convenience
+from . import site, post  # noqa: F401
