+++ app/shell/py/pie/setup.py
@@ -31,8 +31,8 @@ setup(
             'gen-markdown-index=pie.gen_markdown_index:main',
             'check-page-title=pie.check_page_title:main',
             'check-post-build=pie.check_post_build:main',
-            'create-post=pie.create_post:main',
-            'create=pie.create:main',
+            'create-post=pie.create.post:main',
+            'create-site=pie.create.site:main',
             'emojify=pie.emojify:main',
         ],
     },
