+++ app/shell/py/pie/tests/test_create.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from pie import create
+from pie.create import site
 
 
 def test_create_scaffolding(tmp_path: Path) -> None:
     target = tmp_path / "press-project"
-    create.main([str(target)])
+    site.main([str(target)])
 
     assert (target / "docker-compose.yml").exists()
     assert (target / "src").is_dir()
